(PC-13422)[API] refactor: move accessibility validation and status to mixin
