+++ api/src/pcapi/core/offers/repository.py
@@ -24,7 +24,6 @@ from pcapi.core.offers.exceptions import StockDoesNotExist
 from pcapi.core.offers.models import ActivationCode
 from pcapi.core.offers.models import Mediation
 from pcapi.core.offers.models import Offer
-from pcapi.core.offers.models import OfferStatus
 from pcapi.core.offers.models import OfferValidationConfig
 from pcapi.core.offers.models import OfferValidationStatus
 from pcapi.core.offers.models import Stock
@@ -33,6 +32,7 @@ from pcapi.domain.pro_offers.offers_recap import OffersRecap
 from pcapi.infrastructure.repository.pro_offers.offers_recap_domain_converter import to_domain
 from pcapi.models import db
 from pcapi.models.offer_criterion import OfferCriterion
+from pcapi.models.offer_mixin import OfferStatus
 from pcapi.models.product import Product
 from pcapi.models.user_offerer import UserOfferer
 from pcapi.utils.custom_keys import compute_venue_reference
