+++ api/src/pcapi/routes/serialization/offers_serialize.py
@@ -12,9 +12,9 @@ from pcapi.core.bookings.api import compute_cancellation_limit_date
 from pcapi.core.categories.conf import can_create_from_isbn
 from pcapi.core.categories.subcategories import SubcategoryIdEnum
 from pcapi.core.offers import repository as offers_repository
-from pcapi.core.offers.models import OfferStatus
 from pcapi.core.offers.models import Stock
 from pcapi.models.feature import FeatureToggle
+from pcapi.models.offer_mixin import OfferStatus
 from pcapi.routes.serialization import BaseModel
 from pcapi.serialization.utils import dehumanize_field
 from pcapi.serialization.utils import dehumanize_list_field
