+++ api/tests/core/offers/test_models.py
@@ -9,10 +9,10 @@ import pcapi.core.offerers.factories as providers_factories
 from pcapi.core.offers import factories
 from pcapi.core.offers import models
 from pcapi.core.offers.models import Offer
-from pcapi.core.offers.models import OfferStatus
 from pcapi.core.offers.models import OfferValidationStatus
 from pcapi.core.offers.models import Stock
 from pcapi.models.api_errors import ApiErrors
+from pcapi.models.offer_mixin import OfferStatus
 from pcapi.models.pc_object import DeletedRecordException
 from pcapi.repository import repository
 from pcapi.utils.date import DateTimes
